Harden bean purchase value handling

+++ supabase/schemas/dripwatch.sql
@@ -28,7 +28,7 @@ create table if not exists public.beans (
     roast_level text,
     roast_date timestamptz,
     roaster_notes text,
-    price_sgd double precision
+    price_sgd numeric(12, 2)
         constraint beans_price_sgd_positive check (price_sgd is null or price_sgd > 0),
     bag_size_grams double precision
         constraint beans_bag_size_grams_positive check (bag_size_grams is null or bag_size_grams > 0),
